[TM-3336] rename file for clarity
